Fix editing blocks that refer to deleted pages for non-admins
When a page is permanently deleted there may be (custom) blocks that still refer to the page. The page selector widget, in edit mode, blocks editing of that field for anyone without superadmin priviledges. By checking if the page actually exists permissions are not checked and the widget becomes editable.