Fix deletion of external links in the trashcan
When an external link page is aliased, that page become undeletable in the trashcan. It also blocks all other pages from being deleted as soon as it is the first in the queue. This commit fixes that by adding the external link check before entering an infinite loop between `$this->delete()` and `$this->removeThisAlias()`.