Remove dangling CollectionVersionBlocks
It happens from time to time that there is a row in CollectionVersionBlocks referencing a bID to a block which no longer exists. These rows currently do not get deleted when you delete a version. This can "break" the page.
Lets say the newest/live version is 3. Now you decide to approve version 2 and afterwards you delete version 3. Now there is still a row in CollectionVersionBlocks referring to version 3 because the block it referenced does no longer exist. This means you cannot create a new version anymore and instead you will get a "Integrity constraint violation: 1062 Duplicate entry" error.
This is not a perfect solution because it does not address where there is a row referencing a block which no longer exists. But it will clean up "dangling" rows when you delete a version, making sure you do not brick the page.