This adds the possibility to display SVG images in the File Manager.
1. As a detail image in the File Manager's list
2. As a larger Image in the "details" view of the File Manager
3. As an even larger image from the details view replacing the thumbnails view.

Normally the details view gives access to the image thumbnails if any. It wouldn't make sense for SVG files so the option is replaced with a simple "View the SVG file" option instead.

This also fixes a small issue in the "details" view. Files that are not images and cannot have thumbnails such as PDF files were not showing anything. This fixes it to display the Generic Icon for the file type, the same already used in the File Manager's list.